libbpf-rs: Move program definitions into separate file
